requirements.txt
Pillow>=10.3.0
Pillow>=11.1.0
python-dateutil>=2.8.2
tqdm>=4.67.1
piexif>=1.1.3
ffmpeg-python>=0.2.0
pillow-heif==0.21.0
rawpy>=0.18.1
plotly>=5.19.0